lsb: add check before create link
Issue: CGP9-133

* Add the logic to check the existence of the file
  before create a symbolic link for it to avoid
  broken symbolic link

(LOCAL REV: NOT UPSTREAM) -- Send to oe-core 20170303

+++ meta/recipes-extended/lsb/lsb_4.1.bb
@@ -90,9 +90,11 @@ do_install_append(){
        install -m 0755 ${WORKDIR}/init-functions ${D}/lib/lsb
 
        # creat links for LSB test
-       install -d ${D}/usr/lib/lsb
-       ln -sf ${sbindir}/chkconfig ${D}/usr/lib/lsb/install_initd
-       ln -sf ${sbindir}/chkconfig ${D}/usr/lib/lsb/remove_initd
+       if [ -f ${sbindir}/chkconfig ];then
+	       install -d ${D}/usr/lib/lsb
+	       ln -sf ${sbindir}/chkconfig ${D}/usr/lib/lsb/install_initd
+	       ln -sf ${sbindir}/chkconfig ${D}/usr/lib/lsb/remove_initd
+       fi
 
        if [ "${TARGET_ARCH}" = "x86_64" ];then
 	       cd ${D}
